fix quota retrieval in windows systems
The quota logic was using a syscall which is not available in windows.
